+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsGrid.tsx
@@ -8,7 +8,7 @@ import {
   CallParticipantsListComponentProps,
 } from '../CallParticipantsList/CallParticipantsList';
 import { ComponentTestIds } from '../../../constants/TestIds';
-import { theme } from '../../../theme';
+import { useTheme } from '../../../contexts/ThemeContext';
 
 /**
  * Props for the CallParticipantsGrid component.
@@ -32,6 +32,9 @@ export const CallParticipantsGrid = ({
   ParticipantView,
   VideoRenderer,
 }: CallParticipantsGridProps) => {
+  const {
+    theme: { colors, callParticipantsGrid },
+  } = useTheme();
   const { useRemoteParticipants, useParticipants } = useCallStateHooks();
   const _remoteParticipants = useRemoteParticipants();
   const allParticipants = useParticipants();
@@ -53,7 +56,11 @@ export const CallParticipantsGrid = ({
 
   return (
     <View
-      style={styles.container}
+      style={[
+        styles.container,
+        { backgroundColor: colors.dark_gray },
+        callParticipantsGrid.container,
+      ]}
       testID={ComponentTestIds.CALL_PARTICIPANTS_GRID}
     >
       {CallParticipantsList && (
@@ -69,6 +76,5 @@ export const CallParticipantsGrid = ({
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: theme.light.dark_gray,
   },
 });